backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1

# AI/ML - Docling Integration
torch==2.1.2
transformers==4.36.2
docling==0.1.0
langchain==0.1.0
langchain-community==0.0.10

# Enhanced Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10

# Core ML/AI
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.3
scipy==1.11.4

# Phase 3: Enhanced Features
# Search and Vector Database
opensearch-py==2.4.0
qdrant-client==1.7.0

# NLP and Healthcare
scispacy==0.5.3
en-core-sci-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3-py3-none-any.whl

# Phase 4: Advanced Features
# Real-time Notifications and WebSockets
websockets==12.0
redis==5.0.1

# Background Tasks and Message Queuing
celery==5.3.4

# Machine Learning and Analytics
scikit-learn==1.3.2
plotly==5.17.0
dash==2.14.2
pandas-profiling==3.6.6

# Performance and Monitoring
prometheus-client==0.19.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
